# docs/swagger.yaml
basePath: /
definitions:
...
  request.CarUpgradesRequest:
    properties:
      carId:
        type: string
      paymentMode:
        type: string
    type: object
...
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the doumentation for drag racing game
  title: Drag Racing
  version: "2.0"
paths:
...
  /car/upgrade/body:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: body upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade body
      tags:
      - Car-Upgrades
  /car/upgrade/engine:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Engine upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade Engine
      tags:
      - Car-Upgrades
  /car/upgrade/intake:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: intake upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade intake
      tags:
      - Car-Upgrades
  /car/upgrade/nitrous:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: nitrous upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade nitrous
      tags:
      - Car-Upgrades
  /car/upgrade/tires:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: tires upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade tires
      tags:
      - Car-Upgrades
  /car/upgrade/transmission:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: transmission upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade transmission
      tags:
      - Car-Upgrades
  /car/upgrade/turbo:
    put:
      consumes:
      - application/json
      description: Upgrade the engine of a player's car
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Upgrade Request
        in: body
        name: upgradeRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarUpgradesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Turbo upgraded successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Upgrade Turbo
      tags:
      - Car-Upgrades
...
swagger: "2.0"